+++ index.js
@@ -22,6 +22,7 @@
  */
 module.exports =  {
   Digraph: require('graphlib').Digraph,
+  Graph: require('graphlib').Graph,
   Renderer: require('./lib/Renderer'),
   json: require('graphlib').converter.json,
   layout: require('dagre').layout,